Time: 4 ms (14.62%), Space: 13.5 MB (59.52%) - LeetHub

+++ 1574-maximum-product-of-two-elements-in-an-array/1574-maximum-product-of-two-elements-in-an-array.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     int maxProduct(vector<int>& nums) {
         int maxValue = INT_MIN;
-        unsigned short numsSize = nums.size();
+        int numsSize = nums.size();
 
-        for (unsigned short i = 0; i < numsSize - 1; ++i)
-            for (unsigned short j = i + 1; j < numsSize; ++j)
+        for (int i = 0; i < numsSize - 1; ++i)
+            for (int j = i + 1; j < numsSize; ++j)
                 maxValue = max(maxValue, (nums[i] - 1) * (nums[j] - 1));
 
         return maxValue;
